backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic==2.7.1
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose
cryptography
httpx
email-validator
PyNaCl
python-multipart==0.0.9
groq==0.9.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
openpyxl==3.1.2
